+++ go/piqley_plugin_sdk.go
@@ -1,5 +1,0 @@
-package piqley_plugin_sdk
-
-func Hello() string {
-	return "Hello from piqley-plugin-sdk (Go)!"
-}